Added handling of last0

+++ sections/last-software.sh
@@ -30,44 +30,48 @@ function last_software_enforce() {
     # Unpack the WINE directory containing the CME2 utility
     #
 
-    # shellcheck disable=SC2154
-    local wine_dir="${user_home}/.wine"
-    local wine_tgz="${selected_container}/packages/wine+CME2.tgz"
-    message_info "Unpacking the wine+CME2 repository ..."
-    if [ -d "${wine_dir}" ]; then
-        message_success "The directory ${wine_dir} exists"
-    elif [ -r "${wine_tgz}" ]; then
-        if su "${user_last}" -c "cd ~${user_last}; mkdir -p .wine; tar xzf ${wine_tgz}"; then
-            message_success "Extracted ${wine_tgz} into ${wine_dir}"
+    if ! macmap_this_is_last0; then
+        # shellcheck disable=SC2154
+        local wine_dir="${user_home}/.wine"
+        local wine_tgz="${selected_container}/packages/wine+CME2.tgz"
+        message_info "Unpacking the wine+CME2 repository ..."
+        if [ -d "${wine_dir}" ]; then
+            message_success "The directory ${wine_dir} exists"
+        elif [ -r "${wine_tgz}" ]; then
+            if su "${user_last}" -c "cd ~${user_last}; mkdir -p .wine; tar xzf ${wine_tgz}"; then
+                message_success "Extracted ${wine_tgz} into ${wine_dir}"
+            else
+                message_failure "Could not extract ${wine_tgz} into ${wine_dir}"
+            fi
         else
-            message_failure "Could not extract ${wine_tgz} into ${wine_dir}"
+            message_failure "Missing ${wine_tgz}"
         fi
-    else
-        message_failure "Missing ${wine_tgz}"
     fi
 
-    #
-    # Unpack the QFY SDK
-    #
-    local libdir="/usr/local/lib"
-    local package="${selected_container}/packages/sdk_linux64_21.07.16.tgz"
+    if ! macmap_this_is_last0; then
+        #
+        # Unpack the QFY SDK
+        #
+        local libdir="/usr/local/lib"
+        local package="${selected_container}/packages/sdk_linux64_21.07.16.tgz"
 
-    if [ -r "${libdir}/libqhyccd.so.21.7.16.13" ] && [ -L "${libdir}/libqhyccd.so" ] && [ -L "${libdir}/libqhyccd.so.20" ]; then
-        message_success "qhy: The QHY SDK (v21.7.16.13) is installed"
-    elif [ -r "${package}" ]; then
-        local tmp
-        tmp=$(mktemp -d)
+        if [ -r "${libdir}/libqhyccd.so.21.7.16.13" ] && [ -L "${libdir}/libqhyccd.so" ] && [ -L "${libdir}/libqhyccd.so.20" ]; then
+            message_success "qhy: The QHY SDK (v21.7.16.13) is installed"
+        elif [ -r "${package}" ]; then
+            local tmp
+            tmp=$(mktemp -d)
 
-        pushd "${tmp}" >/dev/null 2>&1 || :
-        tar xzf "${package}"
-		cd sdk_linux64_21.07.16 || true
-        chmod +x install.sh
-        ./install.sh
-        popd >/dev/null 2>&1 || :
-        /bin/rm -rf "${tmp}"
-        message_success "Installed the QHY SDK from ${package}"
-    else
-        message_failure "Missing ${package}"
+            pushd "${tmp}" >/dev/null 2>&1 || :
+            tar xzf "${package}"
+            cd sdk_linux64_21.07.16 || true
+            chmod +x install.sh
+            ./install.sh
+            popd >/dev/null 2>&1 || :
+            /bin/rm -rf "${tmp}"
+            message_success "Installed the QHY SDK from ${package}"
+        else
+            message_failure "Missing ${package}"
+        fi
     fi
 
     #
@@ -98,19 +102,21 @@ function last_software_check() {
     su "${user_last}" -c "${fetcher} --dir ~${user_last}/matlab --check"
     (( ret += $? ))
 
-    if [ -d "${wine_dir}" ]; then
-        message_success "The directory ${wine_dir} exists"
-    else
-        message_failure "The ${wine_dir} directory does not exist"
-        (( ret++ ))
-    fi
+    if ! macmap_this_is_last0; then
+        if [ -d "${wine_dir}" ]; then
+            message_success "The directory ${wine_dir} exists"
+        else
+            message_failure "The ${wine_dir} directory does not exist"
+            (( ret++ ))
+        fi
 
-    local libdir="/usr/local/lib"
-    if [ -r "${libdir}/libqhyccd.so.21.7.16.13" ] && [ -L "${libdir}/libqhyccd.so" ] && [ -L "${libdir}/libqhyccd.so.20" ]; then
-        message_success "The QHY SDK (v21.7.16.13) is installed"
-    else
-        message_failure "The QHY SDK (v21.7.16.13) NOT is installed"
-        (( ret++ ))
+        local libdir="/usr/local/lib"
+        if [ -r "${libdir}/libqhyccd.so.21.7.16.13" ] && [ -L "${libdir}/libqhyccd.so" ] && [ -L "${libdir}/libqhyccd.so.20" ]; then
+            message_success "The QHY SDK (v21.7.16.13) is installed"
+        else
+            message_failure "The QHY SDK (v21.7.16.13) NOT is installed"
+            (( ret++ ))
+        fi
     fi
 
     if dpkg -L nomachine >/dev/null 2>&1; then
@@ -146,7 +152,7 @@ EOF
 
     cat <<- EOF
     The following packages cannot be installed from apt repositories, so they get installed from
-     LAST containers:
+     LAST packages:
      - A 'wine' repository for the "Copley Motion" windows software
      - The QHY SDK (v21.7.16.13)
      - Nomachine
